backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
email-validator==2.2.0

pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
aiosqlite==0.20.0
